Update workflow model and serializer for improved metrics tracking
- Changed the `runs_count` field from FloatField to IntegerField in the WorkFlow model to ensure accurate counting of workflow executions.
- Enhanced the WorkFlowSerializer to include detailed logging in the `to_representation` method for better debugging and tracking of workflow state.
- Updated the WorkFlowViewSet and workflow execution service to utilize atomic updates for `last_run` and `runs_count`, improving data integrity during concurrent updates.
- Added a new migration to reflect the changes in the database schema.

+++ apps/workflow/models.py
@@ -1,4 +1,4 @@
-from django.db.models import Model, JSONField, UUIDField, CharField, TextField, DateTimeField, FloatField, ForeignKey, CASCADE, FileField
+from django.db.models import Model, JSONField, UUIDField, CharField, TextField, DateTimeField, FloatField, IntegerField, ForeignKey, CASCADE, FileField
 from django.core.exceptions import ValidationError
 import uuid
 import os
@@ -25,7 +25,7 @@ class WorkFlow(BaseModel):
     ], default='inactive')
     last_run = DateTimeField(blank=True, null=True)
     next_run = DateTimeField(blank=True, null=True)
-    runs_count = FloatField(default=0)
+    runs_count = IntegerField(default=0)
     success_rate = FloatField(default=0)
     tags = JSONField(default=list)
     created_by = CharField(max_length=100, blank=True, null=True)
